resize instance masks with bilinear method

+++ official/vision/beta/projects/panoptic_maskrcnn/dataloaders/panoptic_maskrcnn_input.py
@@ -334,7 +334,9 @@ class Parser(maskrcnn_input.Parser):
       panoptic_instance_mask = panoptic_instance_mask[:, :, 0]
 
       labels['groundtruths'].update({
-          'gt_panoptic_category_mask': panoptic_category_mask,
-          'gt_panoptic_instance_mask': panoptic_instance_mask})
+          'gt_panoptic_category_mask':
+              tf.cast(panoptic_category_mask, dtype=tf.int32),
+          'gt_panoptic_instance_mask':
+              tf.cast(panoptic_instance_mask, dtype=tf.int32)})
 
     return image, labels
